migrate versions test to new test framework
