--- crates/test-rust-wasm/src/bin/versions.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/versions/wasm.rs");
-
-fn main() {}
