+++ tests/runtime/main.rs
@@ -27,7 +27,6 @@ mod resource_into_inner;
 mod resource_with_lists;
 mod resources;
 mod results;
-mod versions;
 
 struct MyCtx {}
